Moved rbs related methods from specs to helper
