+++ spec/integrations/sorbet/rbi_integration_spec.rb
@@ -4,47 +4,6 @@ require 'tmpdir'
 require 'fileutils'
 
 RSpec.describe 'Sorbet RBI integration' do
-  def inline_with_signature_files(code:, rbi:, rbs: nil, rbi_dir_name: 'sorbet/rbi', sig_dir_name: 'sig')
-    skip_unless_sorbet_bridge_available!
-
-    Dir.mktmpdir do |dir|
-      rbi_dir = File.join(dir, rbi_dir_name)
-      FileUtils.mkdir_p(rbi_dir)
-      File.write(File.join(rbi_dir, 'demo.rbi'), rbi)
-
-      raw = {
-        'sorbet' => {
-          'enabled' => true,
-          'rbi_dirs' => [rbi_dir]
-        }
-      }
-
-      if rbs
-        sig_dir = File.join(dir, sig_dir_name)
-        FileUtils.mkdir_p(sig_dir)
-        File.write(File.join(sig_dir, 'demo.rbs'), rbs)
-
-        raw['rbs'] = {
-          'enabled' => true,
-          'sig_dirs' => [sig_dir]
-        }
-      end
-
-      conf = Docscribe::Config.new(raw)
-      Docscribe::InlineRewriter.insert_comments(code, config: conf)
-    end
-  end
-
-  def skip_unless_sorbet_bridge_available!
-    begin
-      require 'rbs'
-    rescue LoadError
-      skip 'RBS not available'
-    end
-
-    skip 'RubyVM::AbstractSyntaxTree not available' unless defined?(RubyVM::AbstractSyntaxTree)
-  end
-
   it 'uses RBI signatures for params and return types' do
     rbi = <<~RBI
       # typed: strict
