+++ spec/spec_helper.rb
@@ -22,6 +22,7 @@ RSpec.configure do |config|
   config.include InlineHelper
   config.include ParamTag
   config.include ExeHelper
+  config.include RbsHelper
   config.example_status_persistence_file_path = '.rspec_status'
   config.disable_monkey_patching!
   config.expect_with(:rspec) { |c| c.syntax = :expect }
